refactor(uzicli): split repo/worker/project-sync verbs into client_repos.go/fake_repos.go (PRD #1017 M3)
Pure motion: move the final 7 HTTPClient verbs (repos, workers, project-sync) and
their FakeClient twins into per-domain files. client.go and fake.go now hold zero
verb methods; ProjectSyncStatus DTO stays with the consts in client.go.
go doc -u -all byte-identical; no test edits.
